working story board

+++ take2/main.js
@@ -3,59 +3,67 @@ let onDocReady = ()=>{
   const headText = document.getElementById('head_text');
   const headDiv = $("#head");
   const headCont = $("#head_cont");
-
-  let chose = [];
-  let story1 = new Story(chose);
-  let typeNext = story1.getStory();
-
-  let callFunc = ()=>{
-    typeNext = story1.getStory()
+  
+  let JamiePullThatUp = ()=>{
+    $("#head_background").css("top","0vh")
+    //typeNext = story1.getStory()
     headCont.css("top","-27.5vh")
     headDiv.toggleClass("user story")
     headDiv.css("animation-name","head-min");
     $("#head_background").css("top","-55vh")
-    for(let i = 0;i<=buttons.length;i++){
-      $(buttons[i]).html(story1.getChoiceInfo("a"))
-    }
+    $(buttons[0]).html(story1.getChoiceInfo("a"))
+    $(buttons[1]).html(story1.getChoiceInfo("b"))
   }
-
-  let JamiePullThatUp = ()=>{
+  let JamiePutThatAway = ()=>{
+    headCont.css("top","0vh")
+    headDiv.toggleClass("user story")
+    headDiv.css("animation-name","head-max");
     $("#head_background").css("top","0vh")
   }
-
-  //Type background part of story
+  let chose = [];
+  let story1 = new Story(chose);
+  let dblClicked = false;
+  let typeNext = story1.getStory();
   let typewriterConfig = {
     strings: back,
     autoStart:true,
     loop:false,
-    delay: 150
+    delay: "natural"
   }
   const typewriter = new Typewriter(headText, typewriterConfig)
-    .callFunction(JamiePullThatUp);
-
-  typewriter.typeString(typeNext)
-    .callFunction(callFunc);
-  
-  function clickedChoice(){
-    JamiePullThatUp();
-    headCont.css("top","0vh")
-    headDiv.toggleClass("user story")
-    headDiv.css("animation-name","head-max");
-
-    let clickedButton = $(this).attr('id');
-    console.log(clickedButton)
-    if(clickedButton === "choice_a"){
+  typewriter.deleteAll()
+    .typeString(typeNext)
+    .callFunction(JamiePullThatUp)
+    .start();
+  let type = ()=>{
+    typeNext = story1.getStory();
+    if(dblClicked){
+      typewriter.pause()
+        .callFunction(JamiePullThatUp)
+        .start();
+      console.log(2)
+      //console.log($(".Typewriter__wrapper").html(typeNext));
+    }else{
+      dblClicked = true;
+      typewriter.deleteAll()
+        .typeString(typeNext)
+        .callFunction(JamiePullThatUp)
+        .start();
+    }
+    
+  }
+  function optionClicked(){
+    let UserClicked = $(this).attr("id");
+    if(UserClicked === "choice_a"){
       story1.addChose("a");
     }else{
       story1.addChose("b");
     }
-    typeNext = story1.getStory();
-    typewriter.deleteAll()
-      .typeString(typeNext)
-      .callFunction(callFunc)
-      .start();
-
-  };
-  buttons.click(clickedChoice)
+    dblClicked = false;
+    JamiePutThatAway();
+    type();
+  }
+  $("#head").click(type);
+  buttons.click(optionClicked)
 }
 $(document).ready(onDocReady);
